Struct super::Config

pub struct Config { /* fields omitted */ }

Config structure.

Contains configuration related fields. It is used for storing the configuration parameters and checking their values. Implements the Default trait.

Methods

impl Config

fn from_cli(cli: ArgMatches<'static>) -> Result<Config, Error>

Creates a new Config struct.

fn check(&self) -> bool

Checks if all the needed folders and files exist.

fn get_errors(&self) -> Vec<String>

Returns the folders and files that do not exist.

fn get_loaded_config_files(&self) -> VecIter<PathBuf>

Returns the currently loaded config files.

fn get_app_packages(&self) -> Vec<PathBuf>

Returns the app package.

fn is_verbose(&self) -> bool

Returns true if the application is running in --verbose mode, false otherwise.

fn is_quiet(&self) -> bool

Returns true if the application is running in --quiet mode, false otherwise.

fn is_force(&self) -> bool

Returns true if the application is running in --force mode, false otherwise.

fn set_force(&mut self)

Sets the application to force recreate the analysis files and results temporarily.

fn reset_force(&mut self)

Resets the --force option, so that it gets reset to the configured force option.

fn is_bench(&self) -> bool

Returns true if the application is running in --bench mode, false otherwise.

fn is_open(&self) -> bool

Returns true if the application is running in --open mode, false otherwise.

fn has_to_generate_json(&self) -> bool

Returns true if the application has to generate result in JSON format.

fn has_to_generate_html(&self) -> bool

Returns true if the application has to generate result in HTML format.

fn get_min_criticality(&self) -> Criticality

Returns the min_criticality field.

fn get_threads(&self) -> u8

Returns the threads field.

fn get_dist_folder(&self) -> &Path

Returns the path to the dist_folder.

fn get_results_folder(&self) -> &Path

Returns the path to the results_folder.

fn get_dex2jar_folder(&self) -> &Path

Returns the path to the dex2jar_folder.

fn get_jd_cmd_file(&self) -> &Path

Returns the path to the jd_cmd_file.

fn get_template_path(&self) -> PathBuf

Gets the path to the template.

fn get_templates_folder(&self) -> &Path

Gets the path to the templates folder.

fn get_template_name(&self) -> &str

Gets the name of the template.

fn get_rules_json(&self) -> &Path

Returns the path to the rules_json.

fn get_unknown_permission_criticality(&self) -> Criticality

Returns the criticality of the unknown_permission field.

fn get_unknown_permission_description(&self) -> &str

Returns the description of the unknown_permission field.

fn get_permissions(&self) -> Iter<PermissionConfig>

Returns the loaded permissions.